fix useEffectを分けて問題を解消
