+++ client/src/hooks/useController.tsx
@@ -1,9 +1,13 @@
 import type { UserId } from 'commonTypesWithClient/branded';
+import type { PlayerModel } from 'commonTypesWithClient/models';
+import { useRouter } from 'next/router';
 import type { MouseEvent, TouchEvent } from 'react';
-import { useRef, useState } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import type { IJoystickUpdateEvent } from 'react-joystick-component/build/lib/Joystick';
 import styles from 'src/pages/controller/index.module.css';
+import { staticPath } from 'src/utils/$path';
 import { apiClient } from 'src/utils/apiClient';
+import { getUserIdFromLocalStorage } from 'src/utils/loginWithLocalStorage';
 import { usePerformanceTimer } from './useTimer';
 
 type MoveTo = {
@@ -11,12 +15,17 @@ type MoveTo = {
   y: number;
 };
 
-export const useController = (userId: UserId) => {
+export const useController = () => {
   //ANCHOR - state
+  const [userId, setUserId] = useState<UserId>('' as UserId);
+  const [playerStatus, setPlayerStatus] = useState<PlayerModel>();
+
   const [moveIntervalId, setMoveIntervalId] = useState<NodeJS.Timeout[]>([]);
   const moveDirection = useRef<MoveTo>({ x: 0, y: 0 });
 
   const [shootIntervalId, setShootIntervalId] = useState<NodeJS.Timeout[]>([]);
+
+  const { startTime, endTime, start, end } = usePerformanceTimer();
 
   const {
     startTime: startTime2,
@@ -32,9 +41,30 @@ export const useController = (userId: UserId) => {
     end: end3,
   } = usePerformanceTimer();
 
+  const router = useRouter();
+
   //ANCHOR - constants
   const MOVE_INTERVAL_TIME = 20;
   const SHOOT_INTERVAL_TIME = 50;
+
+  //ANCHOR - player
+  const getUserId = useCallback(async () => {
+    const localStorageUserId = getUserIdFromLocalStorage();
+    if (!(playerStatus?.isPlaying ?? true)) return;
+    if (localStorageUserId === null) {
+      alert('ログインがまだ行われておりません');
+      return router.push('/login');
+    }
+    setUserId(localStorageUserId);
+  }, [router, playerStatus?.isPlaying]);
+
+  const fetchPlayerStatus = useCallback(async () => {
+    start();
+    const res = await apiClient.player.control.$get({ query: { userId } });
+    end();
+    if (res === null) return;
+    setPlayerStatus(res);
+  }, [userId, start, end]);
 
   //ANCHOR - shoot
   const startShoot = async (e: TouchEvent<HTMLButtonElement> | MouseEvent<HTMLButtonElement>) => {
@@ -52,8 +82,8 @@ export const useController = (userId: UserId) => {
         })
         .then(() => end2());
 
-      // const audio = new Audio(staticPath.sounds.shot_mp3);
-      // audio.play();
+      const audio = new Audio(staticPath.sounds.shot_mp3);
+      audio.play();
     }, SHOOT_INTERVAL_TIME);
     setShootIntervalId((prev) => [...prev, shootIntervalId]);
   };
@@ -97,16 +127,39 @@ export const useController = (userId: UserId) => {
   };
 
   //ANCHOR - effect
+  useEffect(() => {
+    const playerStatusIntervalId = setInterval(() => {
+      fetchPlayerStatus();
+    }, 500);
 
+    return () => {
+      clearInterval(playerStatusIntervalId);
+    };
+  }, [fetchPlayerStatus]);
+
+  useEffect(() => {
+    const userIdIntervalId = setInterval(() => {
+      getUserId();
+    }, 2000);
+
+    return () => {
+      clearInterval(userIdIntervalId);
+    };
+  }, [getUserId]);
+
+  const time1 = startTime - endTime;
   const time2 = startTime2 - endTime2;
   const time3 = startTime3 - endTime3;
+  console.log(userId);
   //ANCHOR - return
   return {
+    playerStatus,
     startMove,
     handleMove,
     stopMove,
     startShoot,
     stopShoot,
+    time1,
     time2,
     time3,
   };
