+++ client/src/pages/controller/index.page.tsx
@@ -1,28 +1,32 @@
-import type { UserId } from 'commonTypesWithClient/branded';
-import type { PlayerModel } from 'commonTypesWithClient/models';
-import { useRouter } from 'next/router';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { Joystick } from 'react-joystick-component';
 import GameClear from 'src/components/GameClear/GameClear';
 import { Traffic } from 'src/components/traffic/traffic';
 import { useController } from 'src/hooks/useController';
-import { usePerformanceTimer } from 'src/hooks/useTimer';
-import { apiClient } from 'src/utils/apiClient';
-import { getUserIdFromLocalStorage, logoutWithLocalStorage } from 'src/utils/loginWithLocalStorage';
+import { logoutWithLocalStorage } from 'src/utils/loginWithLocalStorage';
 import styles from './index.module.css';
 
 const Home = () => {
   //ANCHOR - state
-  const [userId, setUserId] = useState<UserId>('' as UserId);
-  const [playerStatus, setPlayerStatus] = useState<PlayerModel>();
-
-  const { startTime, endTime, start, end } = usePerformanceTimer();
-
   const [windowsize, setWindowsize] = useState<{ width: number; height: number }>({
     width: window.innerWidth,
     height: window.innerHeight,
   });
 
+  //ANCHOR -hooks
+  const {
+    playerStatus,
+    startMove,
+    handleMove,
+    stopMove,
+    startShoot,
+    stopShoot,
+    time1,
+    time2,
+    time3,
+  } = useController();
+
+  //ANCHOR - memo
   const joystickSize = useMemo(() => {
     const aspectRatio = windowsize.width / windowsize.height;
     if (aspectRatio > 3 / 4) {
@@ -31,28 +35,6 @@ const Home = () => {
       return windowsize.width * 0.5 * 0.64;
     }
   }, [windowsize]);
-
-  const router = useRouter();
-
-  //ANCHOR - player
-  const getUserId = useCallback(async () => {
-    const localStorageUserId = getUserIdFromLocalStorage();
-    console.log('a');
-    if (!(playerStatus?.isPlaying ?? true)) return;
-    if (localStorageUserId === null) {
-      alert('ログインがまだ行われておりません');
-      return router.push('/login');
-    }
-    setUserId(localStorageUserId);
-  }, [router, playerStatus?.isPlaying]);
-
-  const fetchPlayerStatus = useCallback(async () => {
-    start();
-    const res = await apiClient.player.control.$get({ query: { userId } });
-    end();
-    if (res === null) return;
-    setPlayerStatus(res);
-  }, [userId, start, end]);
 
   //ANCHOR - effect
   useEffect(() => {
@@ -95,25 +77,7 @@ const Home = () => {
     };
   }, []);
 
-  useEffect(() => {
-    const userIdIntervalId = setInterval(() => {
-      getUserId();
-    }, 2000);
-
-    const playerStatusIntervalId = setInterval(() => {
-      fetchPlayerStatus();
-    }, 500);
-
-    return () => {
-      clearInterval(userIdIntervalId);
-      clearInterval(playerStatusIntervalId);
-    };
-  }, [getUserId, fetchPlayerStatus]);
-
-  //ANCHOR - hooks
-  const { startMove, handleMove, stopMove, startShoot, stopShoot, time2, time3 } =
-    useController(userId);
-
+  //ANCHOR - earlyReturn
   if (!(playerStatus?.isPlaying ?? true)) return <GameClear />;
 
   const logout = () => {
@@ -121,9 +85,6 @@ const Home = () => {
     logoutWithLocalStorage();
   };
 
-  const time1 = startTime - endTime;
-
-  console.log('user', userId);
   return (
     <div className={styles.controller}>
       <Traffic
